post series for user

+++ src/main/java/com/tlongdev/spicio/storage/dao/SequenceDao.java
@@ -6,5 +6,5 @@ package com.tlongdev.spicio.storage.dao;
  */
 public interface SequenceDao {
 
-    int nextValue(String id);
+    Long nextValue(String id);
 }
